# api/openapi.yaml
openapi: 3.0.3
info:
  title: OpenTrust API
  version: 0.1.0
paths:
  /api/v1/health:
    get:
      responses:
        "200": { description: OK }
  /api/v1/tools:
    get:
      responses:
        "200": { description: List passports }
    post:
      responses:
        "201": { description: Create passport }
  /api/v1/tools/{slug}:
    get:
      parameters: [{ name: slug, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Get passport }
    put:
      parameters: [{ name: slug, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Update passport }
  /api/v1/tools/{slug}/badge:
    get:
      parameters: [{ name: slug, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Badge location }
  /api/v1/search:
    get:
      parameters: [{ name: q, in: query, required: true, schema: { type: string } }]
      responses:
        "200": { description: Search passports }
  /api/v1/claim:
    post:
      responses:
        "200": { description: Start stateless GitHub OAuth claim }
  /api/v1/claim/callback:
    get:
      responses:
        "200": { description: Return JWT claim token }
  /api/v1/payments/checkout:
    post:
      responses:
        "200": { description: Create demo checkout }
  /api/v1/payments/verify:
    post:
      responses:
        "200": { description: Verify demo checkout }
        "404": { description: Checkout not found }
  /api/v1/subscriptions/create:
    post:
      responses:
        "200": { description: Create demo subscription checkout }
  /api/v1/escrow/create:
    post:
      responses:
        "200": { description: Create trust-verified escrow and return deposit instructions }
        "403": { description: Escrow disabled or seller trust gate failed }
        "422": { description: Listing is missing delivery proof }
  /api/v1/escrow/{escrow_id}:
    get:
      responses:
        "200": { description: Return escrow state }
        "404": { description: Escrow not found }
  /api/v1/escrow/{escrow_id}/verify-deposit:
    post:
      responses:
        "200": { description: Verify Base USDC deposit and mark escrow funded }
        "402": { description: On-chain payment verification failed }
        "404": { description: Escrow or wallet not found }
        "409": { description: Invalid escrow state transition }
  /api/v1/escrow/{escrow_id}/deliver:
    post:
      responses:
        "200": { description: Record delivery proof }
        "404": { description: Escrow not found }
        "409": { description: Escrow is not funded or proof is incomplete }
  /api/v1/escrow/{escrow_id}/release:
    post:
      responses:
        "200": { description: Release escrow through the configured provider }
        "404": { description: Escrow not found }
        "409": { description: Escrow is not deliverable for release }
  /api/v1/escrow/{escrow_id}/refund:
    post:
      responses:
        "200": { description: Refund escrow through the configured provider }
        "404": { description: Escrow not found }
        "409": { description: Escrow is not refundable }
  /api/v1/escrow/{escrow_id}/disputes:
    post:
      responses:
        "200": { description: Open a dispute and block release }
        "404": { description: Escrow not found }
        "409": { description: Escrow is not disputable }
  /api/v1/escrow/{escrow_id}/ratings:
    post:
      parameters: [{ name: escrow_id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Submit a bidirectional counterparty rating after settlement }
        "409": { description: Escrow not settled or already rated by this party }
    get:
      parameters: [{ name: escrow_id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: List ratings submitted for this escrow }
  /api/v1/reputation/{subject_id}:
    get:
      parameters: [{ name: subject_id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Registry-computed reputation record for a client, server, or agent }
        "404": { description: No reputation record for this subject }
  /api/v1/reputation/{subject_id}/ratings:
    get:
      parameters: [{ name: subject_id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Ratings received by this subject }
  /api/v1/jobs:
    post:
      responses:
        "200": { description: Client posts work wanted }
    get:
      responses:
        "200": { description: List jobs with optional status/provider_kind/max_budget filters }
  /api/v1/jobs/{job_id}:
    get:
      parameters: [{ name: job_id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Read one job posting }
  /api/v1/jobs/{job_id}/engage:
    post:
      parameters: [{ name: job_id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Provider engages an open job, minting an escrow }
        "403": { description: Escrow disabled or provider below reputation floor }
  /api/v1/jobs/{job_id}/cancel:
    post:
      parameters: [{ name: job_id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: Client cancels an open job }
        "409": { description: Job is not open }